Day7 and 9

+++ src/main/java/com/gabrielsson/aoc/Day7.java
@@ -1,14 +1,13 @@
 package com.gabrielsson.aoc;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.OptionalInt;
 import java.util.concurrent.ArrayBlockingQueue;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.IntStream;
+import java.util.stream.LongStream;
 
 import org.paukov.combinatorics3.Generator;
 
@@ -18,27 +17,27 @@ public class Day7 {
 
     private static final boolean USE_FEEDBACK = true;
 
-    public static Object part2(List<Integer> listOfRows) {
+    public static Object part2(List<Long> listOfRows) {
         return Generator.permutation(5, 6, 7, 8, 9).simple().stream()
-                .mapToInt(phases -> findLargestThrust(listOfRows, phases)).max().orElse(-1);
+                .mapToLong(phases -> findLargestThrust(listOfRows, phases)).max().orElse(-1);
 
     }
 
-    private static Integer findLargestThrust(List<Integer> program, List<Integer> phases) {
+    private static Long findLargestThrust(List<Long> program, List<Integer> phases) {
         try {
             ThreadPoolExecutor executor = (ThreadPoolExecutor) Executors.newFixedThreadPool(5);
-            List<BlockingQueue<Integer>> wires = IntStream.rangeClosed(0, phases.size())
-                    .mapToObj(c -> new ArrayBlockingQueue<Integer>(2))
+            List<BlockingQueue<Long>> wires = LongStream.rangeClosed(0, phases.size())
+                    .mapToObj(c -> new ArrayBlockingQueue<Long>(2))
                     .collect(toList());
             if (USE_FEEDBACK) {
                 wires.remove(wires.size() - 1);
                 wires.add(wires.size(), wires.get(0));
             }
             for (int i = 0; i < phases.size(); i++) {
-                wires.get(i).put(phases.get(i));
+                wires.get(i).put(Long.valueOf(phases.get(i)));
                 executor.execute(new IntComp(program, wires.get(i), wires.get(i + 1)));
             }
-            wires.get(0).put(0); // Initial Input
+            wires.get(0).put(0l); // Initial Input
             executor.shutdown();
             executor.awaitTermination(1L, TimeUnit.DAYS);
             return wires.get(wires.size() - 1).poll(1, TimeUnit.HOURS);
